feat: update demo sanbox
